+++ app/Http/Controllers/ThirdPartiesController.php
@@ -6,6 +6,11 @@ use App\ThirdParty;
 
 class ThirdPartiesController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+    
     public function render()
     {
         return view('third_parties.index');
@@ -13,6 +18,8 @@ class ThirdPartiesController extends Controller
 
     public function index()
     {
+        $this->authorize('view', ThirdParty::class);
+
         return ThirdParty::orderBy('name')->paginate(30);
     }
 
@@ -23,6 +30,8 @@ class ThirdPartiesController extends Controller
 
     public function store()
     {
+        $this->authorize('create', ThirdParty::class);
+
         $attributes = request()->validate([
             'identification' => 'required|min:8|unique:third_parties,identification',
             'name' => 'required|string|min:3',
@@ -40,6 +49,8 @@ class ThirdPartiesController extends Controller
     public function update(ThirdParty $thirdParty)
     {
 
+        $this->authorize('update', $thirdParty);
+
         $attributes = request()->validate([
             'identification' => 'required|min:8|unique:third_parties,identification,' . $thirdParty->id,
             'name' => 'required|string|min:3',
@@ -56,6 +67,8 @@ class ThirdPartiesController extends Controller
 
     public function destroy(ThirdParty $thirdParty)
     {
+        $this->authorize('delete', $thirdParty);
+
         $thirdParty->delete();
 
         return response()->json(['message' => 'Tercero eliminado correctamente!'], 200);
